Created GenericDirectoryScreen, removed GenericContainer Improved EventsScreen code structure
